requirements.txt
fastapi
python-multipart
uvicorn
hatchet-sdk
python-dotenv
docling
chonkie
tokenizers 